Finalização Domain de Jornada

+++ src/main/java/geo/track/jornada/response/listagem/CardOrdemDeServicoResponse.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 public record CardOrdemDeServicoResponse(
         Integer idOrdemServico,
         Double valorTotal,
-        Integer diasEspera,
+        Long diasEspera,
         LocalDate dataSaidaPrevista,
         LocalDate dataSaidaEfetiva,
         LocalDate dataEntradaPrevista,
